+++ lib/src/credentials/credential_signer.dart
@@ -4,10 +4,13 @@ import 'dart:typed_data';
 
 import 'package:base_codecs/base_codecs.dart';
 import 'package:crypto/crypto.dart';
+import 'package:dart_ssi/src/credentials/jsonLdContext/ecdsa_recovery_2020.dart';
+import 'package:dart_ssi/src/credentials/jsonLdContext/ed25519_Signature.dart';
 import 'package:dart_ssi/src/util/utils.dart';
 import 'package:dart_web3/credentials.dart';
 import 'package:dart_web3/crypto.dart' as web3Crypto;
 import 'package:ed25519_edwards/ed25519_edwards.dart' as ed;
+import 'package:json_ld_processor/json_ld_processor.dart';
 
 import '../wallet/wallet_store.dart';
 import 'credential_operations.dart';
@@ -19,10 +22,11 @@ abstract class Signer {
       {String? challenge, String? domain});
   FutureOr<Uint8List> sign(dynamic data, WalletStore wallet, String did,
       {String? challenge, String? domain});
-  bool verify(dynamic proof, dynamic data, String did, {String? challenge});
+  FutureOr<bool> verify(dynamic proof, dynamic data, String did,
+      {String? challenge});
 }
 
-class EcdsaRecoverySignature extends Signer {
+class EcdsaRecoverySignatureOld extends Signer {
   final String typeName = 'EcdsaSecp256k1RecoverySignature2020';
 
   @override
@@ -148,7 +152,153 @@ class EcdsaRecoverySignature extends Signer {
   }
 }
 
-class EdDsaSigner extends Signer {
+class EcdsaRecoverySignature extends Signer {
+  final String typeName = 'EcdsaSecp256k1RecoverySignature2020';
+  final Function(Uri url, LoadDocumentOptions? options) loadDocument;
+  EcdsaRecoverySignature(this.loadDocument);
+
+  @override
+  Future<Map<String, dynamic>> buildProof(data, WalletStore wallet, String did,
+      {String? challenge, String? domain}) async {
+    var proofOptions = {
+      '@context': ECDSA_RECOVERY_CONTEXT_IRI,
+      'type': typeName,
+      'proofPurpose': 'assertionMethod',
+      'verificationMethod': did,
+      'created': DateTime.now().toUtc().toIso8601String()
+    };
+    if (domain != null) {
+      proofOptions['domain'] = domain;
+    }
+    if (challenge != null) {
+      proofOptions['challenge'] = challenge;
+    }
+
+    List<int> hashToSign = await _dataToHash(data);
+
+    var pOptionsHash = sha256
+        .convert(utf8.encode(await JsonLdProcessor.normalize(proofOptions,
+            options:
+                JsonLdOptions(safeMode: true, documentLoader: loadDocument))))
+        .bytes;
+    var hash = sha256.convert(pOptionsHash + hashToSign).bytes;
+
+    proofOptions.remove('@context');
+
+    var privateKeyHex = await wallet.getPrivateKeyForCredentialDid(did);
+    if (privateKeyHex == null)
+      privateKeyHex = await wallet.getPrivateKeyForConnectionDid(did);
+    if (privateKeyHex == null) throw Exception('Could not find a private key');
+    var key = EthPrivateKey.fromHex(privateKeyHex);
+
+    var sigArray = _buildSignatureArray(Uint8List.fromList(hash), key);
+    while (sigArray.length != 65) {
+      sigArray = _buildSignatureArray(Uint8List.fromList(hash), key);
+    }
+
+    var critical = new Map<String, dynamic>();
+    critical['b64'] = false;
+    proofOptions['jws'] = '${buildJwsHeader(alg: 'ES256K-R', extra: critical)}.'
+        '.${base64UrlEncode(sigArray)}';
+
+    return proofOptions;
+  }
+
+  FutureOr<List<int>> _dataToHash(dynamic data) async {
+    if (data is Uint8List)
+      return data.toList();
+    else if (data is List<int>)
+      return data;
+    else if (data is Map<String, dynamic>) {
+      return sha256
+          .convert(utf8.encode(await JsonLdProcessor.normalize(
+              Map<String, dynamic>.from(data),
+              options:
+                  JsonLdOptions(safeMode: true, documentLoader: loadDocument))))
+          .bytes;
+    } else if (data is String) {
+      return sha256.convert(utf8.encode(data)).bytes;
+    } else {
+      throw Exception('Unknown datatype for data');
+    }
+  }
+
+  @override
+  Uint8List sign(data, WalletStore wallet, String did,
+      {String? challenge, String? domain}) {
+    // TODO: implement sign
+    throw UnimplementedError();
+  }
+
+  List<int> _buildSignatureArray(Uint8List hash, EthPrivateKey privateKey) {
+    web3Crypto.MsgSignature signature =
+        web3Crypto.sign(hash, privateKey.privateKey);
+    List<int> rList = web3Crypto.unsignedIntToBytes(signature.r);
+    if (rList.length < 32) {
+      List<int> rPad = List.filled(32 - rList.length, 0);
+      rList = rPad + rList;
+    }
+    List<int> sList = web3Crypto.unsignedIntToBytes(signature.s);
+    if (sList.length < 32) {
+      List<int> sPad = List.filled(32 - sList.length, 0);
+      sList = sPad + sList;
+    }
+    List<int> sigArray = rList + sList + [signature.v - 27];
+    return sigArray;
+  }
+
+  @override
+  Future<bool> verify(proof, data, String did, {String? challenge}) async {
+    //compare challenge
+    if (challenge != null) {
+      var containingChallenge = proof['challenge'];
+      if (containingChallenge == null)
+        throw Exception('Expected challenge in this credential');
+      if (containingChallenge != challenge)
+        throw Exception('a challenge do not match expected challenge');
+    }
+
+    //verify signature
+    var signature = _getSignatureFromJws(proof['jws']);
+
+    List<int> hash = await _dataToHash(data);
+
+    var jws = proof.remove('jws');
+    proof['@context'] = ECDSA_RECOVERY_CONTEXT_IRI;
+    var proofHash = sha256
+        .convert(utf8.encode(await JsonLdProcessor.normalize(proof,
+            options:
+                JsonLdOptions(safeMode: true, documentLoader: loadDocument))))
+        .bytes;
+    var hashToSign = sha256.convert(proofHash + hash).bytes;
+
+    proof['jws'] = jws;
+    proof.remove('@context');
+
+    var pubKey = web3Crypto.ecRecover(hashToSign as Uint8List, signature);
+
+    var givenAddress = EthereumAddress.fromHex(did.split(':').last);
+
+    return EthereumAddress.fromPublicKey(pubKey).hexEip55 ==
+        givenAddress.hexEip55;
+  }
+
+  web3Crypto.MsgSignature _getSignatureFromJws(String jws) {
+    var splitJws = jws.split('.');
+    Map<String, dynamic> header =
+        jsonDecode(utf8.decode(base64Decode(addPaddingToBase64(splitJws[0]))));
+    if (header['alg'] != 'ES256K-R')
+      throw Exception('Unsupported signature Algorithm ${header['alg']}');
+    var sigArray = base64Decode(addPaddingToBase64(splitJws[2]));
+    if (sigArray.length != 65) throw Exception('wrong signature-length');
+    return new web3Crypto.MsgSignature(
+        web3Crypto.bytesToUnsignedInt(sigArray.sublist(0, 32)),
+        web3Crypto.bytesToUnsignedInt(sigArray.sublist(32, 64)),
+        sigArray[64] + 27);
+  }
+}
+
+class EdDsaSignerOld extends Signer {
   final String typeName = 'Ed25519Signature2020';
   @override
   FutureOr<Map<String, dynamic>> buildProof(
@@ -223,6 +373,109 @@ class EdDsaSigner extends Signer {
     var proofHash = sha256.convert(utf8.encode(jsonEncode(proof))).bytes;
     var hashToSign = sha256.convert(proofHash + hash).bytes;
 
+    proof['proofValue'] = proofValue;
+
+    var encodedKey = did.split(':')[2];
+    var base58DecodedKey = base58BitcoinDecode(encodedKey.substring(1));
+    return ed.verify(
+        ed.PublicKey(base58DecodedKey.sublist(2)),
+        Uint8List.fromList(hashToSign),
+        base58BitcoinDecode(proofValue.substring(1)));
+  }
+}
+
+class EdDsaSigner extends Signer {
+  final String typeName = 'Ed25519Signature2020';
+  final Function(Uri url, LoadDocumentOptions? options) loadDocument;
+  EdDsaSigner(this.loadDocument);
+
+  @override
+  FutureOr<Map<String, dynamic>> buildProof(
+      data, WalletStore wallet, String did,
+      {String? challenge, String? domain}) async {
+    var proofOptions = {
+      '@context': ED25519_SIGNATURE_CONTEXT_IRI,
+      'type': typeName,
+      'proofPurpose': 'assertionMethod',
+      'verificationMethod': did,
+      'created': DateTime.now().toUtc().toIso8601String()
+    };
+    if (domain != null) {
+      proofOptions['domain'] = domain;
+    }
+    if (challenge != null) {
+      proofOptions['challenge'] = challenge;
+    }
+    String pOptions = await JsonLdProcessor.normalize(proofOptions,
+        options: JsonLdOptions(safeMode: true, documentLoader: loadDocument));
+
+    proofOptions.remove('@context');
+
+    List<int> hashToSign = await _dataToHash(data);
+
+    var pOptionsHash = sha256.convert(utf8.encode(pOptions)).bytes;
+    var hash = pOptionsHash + hashToSign;
+
+    var privateKey = await wallet.getPrivateKeyForCredentialDid(did);
+    if (privateKey == null)
+      privateKey = await wallet.getPrivateKeyForConnectionDid(did);
+    if (privateKey == null) throw Exception('Could not find a private key');
+    var signature = ed.sign(
+        ed.PrivateKey(web3Crypto.hexToBytes(privateKey).toList()),
+        Uint8List.fromList(hash));
+
+    proofOptions['proofValue'] = 'z${base58BitcoinEncode(signature)}';
+
+    return proofOptions;
+  }
+
+  FutureOr<List<int>> _dataToHash(dynamic data) async {
+    if (data is Uint8List)
+      return data;
+    else if (data is Map<String, dynamic>) {
+      print(data);
+      var normal = await JsonLdProcessor.normalize(
+          Map<String, dynamic>.from(data),
+          options: JsonLdOptions(safeMode: true, documentLoader: loadDocument));
+      return sha256.convert(utf8.encode(normal)).bytes;
+    } else if (data is String) {
+      return sha256.convert(utf8.encode(data)).bytes;
+    } else {
+      throw Exception('Unknown datatype for data');
+    }
+  }
+
+  @override
+  FutureOr<Uint8List> sign(data, WalletStore wallet, String did,
+      {String? challenge, String? domain}) {
+    // TODO: implement sign
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<bool> verify(proof, data, String did, {String? challenge}) async {
+    //compare challenge
+    if (challenge != null) {
+      var containingChallenge = proof['challenge'];
+      if (containingChallenge == null)
+        throw Exception('Expected challenge in this credential');
+      if (containingChallenge != challenge)
+        throw Exception(
+            'challenge in credential do not match expected challenge');
+    }
+    var proofValue = proof.remove('proofValue');
+    proof['@context'] = ED25519_SIGNATURE_CONTEXT_IRI;
+
+    List<int> hash = await _dataToHash(data);
+
+    var proofHash = sha256
+        .convert(utf8.encode(await JsonLdProcessor.normalize(proof,
+            options:
+                JsonLdOptions(safeMode: true, documentLoader: loadDocument))))
+        .bytes;
+    var hashToSign = proofHash + hash;
+
+    proof.remove('@context');
     proof['proofValue'] = proofValue;
 
     var encodedKey = did.split(':')[2];